# opt: optimize FuncDepSet.inClosureOf