#### opt/bench: add slow-query-5

This commit adds the new `slow-query-5` benchmark which tests a 40+ -way
join modelled off of a real-world production query. The query stresses
the efficiency of the optimizer, particularly in building and
manipulating thousands of functional dependencies for the thousands of
possible join orderings.

Release note: None

#### opt: optimize FuncDepSet.inClosureOf

`FuncDepSet.inClosureOf` is used frequently to manipulate a
`FuncDepSet`. In queries with many tables, columns, and joins, it can
contribute significantly to the latency of query planning, due to the
`O(n²)` time complexity of calculating a transitive closure, where `n`
is the number of dependencies in the set, and the relatively expensive
`SubsetOf` set operation.

This commit optimizes the function by reducing the `O(n²)` time
complexity by some constant value in some cases.

Epic: None

Release note (performance improvement): Query planning time has been
reduced significantly for some queries in which many tables are joined.