@mgartner 2023-10-27
I'll post benchmark results shortly.