@mgartner 2023-10-27
```
name                         old time/op    new time/op    delta
SlowQueries/slow-query-1-24    29.3ms ± 2%    28.8ms ± 0%   -1.61%  (p=0.016 n=5+4)
SlowQueries/slow-query-2-24     413ms ± 1%     408ms ± 1%   -1.38%  (p=0.016 n=5+5)
SlowQueries/slow-query-3-24    97.0ms ± 0%    96.9ms ± 0%     ~     (p=0.548 n=5+5)
SlowQueries/slow-query-4-24     687ms ± 0%     697ms ± 1%   +1.45%  (p=0.008 n=5+5)
SlowQueries/slow-query-5-24     4.34s ± 0%     2.67s ± 1%  -38.46%  (p=0.008 n=5+5)

name                         old alloc/op   new alloc/op   delta
SlowQueries/slow-query-1-24    6.83MB ± 0%    6.89MB ± 0%   +0.95%  (p=0.008 n=5+5)
SlowQueries/slow-query-2-24    48.6MB ± 0%    48.6MB ± 0%     ~     (p=0.841 n=5+5)
SlowQueries/slow-query-3-24    44.8MB ± 0%    45.1MB ± 0%   +0.67%  (p=0.008 n=5+5)
SlowQueries/slow-query-4-24     153MB ± 0%     155MB ± 0%   +1.07%  (p=0.008 n=5+5)
SlowQueries/slow-query-5-24     405MB ± 0%     522MB ± 0%  +28.75%  (p=0.029 n=4+4)

name                         old allocs/op  new allocs/op  delta
SlowQueries/slow-query-1-24     70.2k ± 0%     72.2k ± 0%   +2.91%  (p=0.008 n=5+5)
SlowQueries/slow-query-2-24      391k ± 0%      391k ± 0%     ~     (p=0.548 n=5+5)
SlowQueries/slow-query-3-24      389k ± 0%      399k ± 0%   +2.41%  (p=0.008 n=5+5)
SlowQueries/slow-query-4-24     3.23M ± 0%     3.28M ± 0%   +1.59%  (p=0.008 n=5+5)
SlowQueries/slow-query-5-24     8.94M ± 0%    12.58M ± 0%  +40.74%  (p=0.008 n=5+5)
```

```
name                                                    old time/op    new time/op    delta
Phases/kv-read/Simple/Explore-24                          49.7µs ± 0%    49.9µs ± 0%    ~     (p=0.095 n=5+5)
Phases/kv-read-const/Simple/Explore-24                    50.4µs ± 0%    50.8µs ± 0%  +0.77%  (p=0.016 n=5+5)
Phases/tpcc-new-order/Simple/Explore-24                    134µs ± 0%     134µs ± 0%    ~     (p=0.421 n=5+5)
Phases/tpcc-delivery/Simple/Explore-24                     115µs ± 1%     115µs ± 1%    ~     (p=0.222 n=5+5)
Phases/tpcc-delivery/Prepared/Explore-24                  43.2µs ± 1%    43.3µs ± 1%    ~     (p=0.841 n=5+5)
Phases/tpcc-stock-level/Simple/Explore-24                  608µs ± 0%     606µs ± 0%    ~     (p=0.222 n=5+5)
Phases/tpcc-stock-level/Prepared/Explore-24                263µs ± 0%     262µs ± 0%  -0.32%  (p=0.032 n=5+5)
Phases/many-columns-and-indexes-a/Simple/Explore-24        455µs ± 0%     463µs ± 0%  +1.80%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/Explore-24      403µs ± 0%     408µs ± 0%  +1.28%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Simple/Explore-24        544µs ± 0%     552µs ± 0%  +1.53%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/Explore-24      458µs ± 0%     464µs ± 0%  +1.34%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Simple/Explore-24        335ms ± 1%     327ms ± 1%  -2.34%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24      341ms ± 1%     332ms ± 1%  -2.51%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-24        138µs ± 0%     137µs ± 0%  -0.44%  (p=0.032 n=5+5)
Phases/single-col-histogram-range/Prepared/Explore-24     58.9µs ± 0%    59.0µs ± 1%    ~     (p=0.222 n=5+5)
Phases/batch-insert-one/Simple/Explore-24                  119µs ± 1%     119µs ± 0%    ~     (p=0.151 n=5+5)
Phases/batch-insert-many/Simple/Explore-24                8.61ms ± 1%    8.76ms ± 1%  +1.73%  (p=0.032 n=5+5)
Phases/ored-preds-100/Simple/Explore-24                   3.68ms ± 0%    3.67ms ± 1%    ~     (p=0.421 n=5+5)
Phases/ored-preds-using-params-100/Simple/Explore-24      3.51ms ± 1%    3.49ms ± 0%    ~     (p=0.056 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24    1.56ms ± 1%    1.53ms ± 1%  -2.10%  (p=0.008 n=5+5)

name                                                    old alloc/op   new alloc/op   delta
Phases/kv-read/Simple/Explore-24                          15.4kB ± 0%    15.4kB ± 0%    ~     (p=0.444 n=5+5)
Phases/kv-read-const/Simple/Explore-24                    15.5kB ± 0%    15.5kB ± 0%    ~     (all equal)
Phases/tpcc-new-order/Simple/Explore-24                   32.7kB ± 0%    32.7kB ± 0%    ~     (p=0.222 n=5+5)
Phases/tpcc-delivery/Simple/Explore-24                    26.7kB ± 0%    26.7kB ± 0%    ~     (p=0.143 n=5+5)
Phases/tpcc-delivery/Prepared/Explore-24                  12.8kB ± 0%    12.8kB ± 0%    ~     (p=1.000 n=5+5)
Phases/tpcc-stock-level/Simple/Explore-24                  130kB ± 0%     130kB ± 0%    ~     (p=0.810 n=5+5)
Phases/tpcc-stock-level/Prepared/Explore-24               58.9kB ± 0%    58.9kB ± 0%    ~     (p=0.071 n=5+5)
Phases/many-columns-and-indexes-a/Simple/Explore-24       39.7kB ± 0%    39.7kB ± 0%    ~     (p=0.984 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/Explore-24     5.50kB ± 0%    5.50kB ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-b/Simple/Explore-24       48.0kB ± 0%    48.0kB ± 0%    ~     (p=0.984 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/Explore-24     8.02kB ± 0%    8.02kB ± 0%    ~     (p=1.000 n=5+5)
Phases/many-columns-and-indexes-c/Simple/Explore-24        110MB ± 0%     111MB ± 0%  +0.95%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24      109MB ± 0%     110MB ± 0%  +0.94%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-24       29.8kB ± 0%    29.8kB ± 0%    ~     (p=0.992 n=5+5)
Phases/single-col-histogram-range/Prepared/Explore-24     13.0kB ± 0%    13.0kB ± 0%    ~     (all equal)
Phases/batch-insert-one/Simple/Explore-24                 64.8kB ± 0%    64.8kB ± 0%    ~     (p=0.167 n=5+5)
Phases/batch-insert-many/Simple/Explore-24                3.92MB ± 0%    3.92MB ± 0%    ~     (p=1.000 n=5+5)
Phases/ored-preds-100/Simple/Explore-24                    945kB ± 0%     945kB ± 0%  -0.01%  (p=0.040 n=5+5)
Phases/ored-preds-using-params-100/Simple/Explore-24       926kB ± 0%     926kB ± 0%    ~     (p=0.690 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24     736kB ± 0%     736kB ± 0%    ~     (p=0.548 n=5+5)

name                                                    old allocs/op  new allocs/op  delta
Phases/kv-read/Simple/Explore-24                            97.0 ± 0%      97.0 ± 0%    ~     (all equal)
Phases/kv-read-const/Simple/Explore-24                      97.0 ± 0%      97.0 ± 0%    ~     (all equal)
Phases/tpcc-new-order/Simple/Explore-24                      177 ± 0%       177 ± 0%    ~     (all equal)
Phases/tpcc-delivery/Simple/Explore-24                       167 ± 0%       167 ± 0%    ~     (all equal)
Phases/tpcc-delivery/Prepared/Explore-24                    75.0 ± 0%      75.0 ± 0%    ~     (all equal)
Phases/tpcc-stock-level/Simple/Explore-24                    711 ± 0%       711 ± 0%    ~     (all equal)
Phases/tpcc-stock-level/Prepared/Explore-24                  341 ± 0%       341 ± 0%    ~     (p=0.444 n=5+5)
Phases/many-columns-and-indexes-a/Simple/Explore-24         83.0 ± 0%      83.0 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-a/Prepared/Explore-24       31.0 ± 0%      31.0 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-b/Simple/Explore-24          138 ± 0%       138 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-b/Prepared/Explore-24       49.0 ± 0%      49.0 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-c/Simple/Explore-24         762k ± 0%      795k ± 0%  +4.28%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24       756k ± 0%      788k ± 0%  +4.22%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-24          524 ± 0%       524 ± 0%    ~     (all equal)
Phases/single-col-histogram-range/Prepared/Explore-24        251 ± 0%       251 ± 0%    ~     (all equal)
Phases/batch-insert-one/Simple/Explore-24                    325 ± 0%       325 ± 0%    ~     (all equal)
Phases/batch-insert-many/Simple/Explore-24                 28.5k ± 0%     28.5k ± 0%    ~     (p=1.000 n=5+4)
Phases/ored-preds-100/Simple/Explore-24                    4.53k ± 0%     4.53k ± 0%  -0.02%  (p=0.029 n=4+4)
Phases/ored-preds-using-params-100/Simple/Explore-24       4.53k ± 0%     4.53k ± 0%    ~     (p=1.000 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24     2.88k ± 0%     2.88k ± 0%    ~     (all equal)
```

Benchmarks were run on my GCE worker.